requirements.txt
alembic==0.9.1
attrs==16.3.0
corral-pipeline
coverage==4.3.4
flake8==2.6.2
funcsigs==1.0.2
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.5.3
mock==2.0.0
pbr==2.0.0
pep8==1.7.0
pycodestyle==2.0.0
pyflakes==1.2.3
python-editor==1.0.3
sadisplay==0.4.6
sh==1.12.11
six==1.10.0
SQLAlchemy==1.1.6
SQLAlchemy-Utils==0.32.12
termcolor==1.1.0
texttable==0.8.7
xmltodict==0.10.2
